+++ Backend/routes/routes.py
@@ -3,28 +3,25 @@ from Backend.config.database import SupplementaryCollection
 from typing import List
 from bson import ObjectId
 from fastapi import APIRouter, HTTPException
-from Backend.model.model import Ngo, User, Donor, Aanganwadi, Child, ChildMalnutrition, SupplementsDetail, Program, \
+from Backend.model.model import Ngo, User, Donor, Child, ChildMalnutrition, SupplementsDetail, Program, \
     ProgramJoining
 from Backend.config.database import UserCollection, AanganwadiCollection, ChildCollection
 from Backend.config.database import DonorsCollection, ChildMalnutritionCollection, SupplementDetailsCollection, \
     ProgramsCollection, ProgramJoiningCollection, AangawadiSummaryCollection
-from Backend.schemas.schema import user_list_serializer, donors_list_serializer, \
-    program_list_serializer
+from Backend.schemas.schema import user_list_serializer, donors_list_serializer, program_list_serializer
 from Backend.schemas.schema import supplements_list_serializer
-from Backend.schemas.schema import aanganwadi_list_serializer, child_list_serializer, \
+from Backend.schemas.schema import child_list_serializer, \
     child_malnutrition_list_serializer, programjoining_list_serializer, supplementaryPacks_list_serializer
 from typing import Union
 
 user_router = APIRouter()
 ngo_router = APIRouter()
 donor_router = APIRouter()
-aanganwadi_router = APIRouter()
 child_router = APIRouter()
 child_malnutrition = APIRouter()
 supplement_details = APIRouter()
 program_router = APIRouter()
 program_joining = APIRouter()
-
 aaganwadi_summary = APIRouter()
 
 
@@ -103,40 +100,6 @@ async def update_donor(id: str, donor: Donor):
     return {"status": "ok", "data": updated_data}
 
 
-@aanganwadi_router.post("/api/addAanganwadi")
-async def aanganwadi_addition(aanganwadi: Aanganwadi):
-    _id = AanganwadiCollection.insert_one(dict(aanganwadi))
-    added_aanganwadi = aanganwadi_list_serializer(
-        AanganwadiCollection.find({"_id": _id.inserted_id}))
-    return {"status": "ok", "data": added_aanganwadi}
-
-
-@aanganwadi_router.get("/api/getAanganwadis")
-async def get_aanganwadis():
-    aanganwadis = aanganwadi_list_serializer(AanganwadiCollection.find())
-    return {"status": "ok", "data": aanganwadis}
-
-
-@aanganwadi_router.get("/api/{id}/get_aanganwadi")
-async def get_aanganwadi(id: str):
-    aanganwadi = aanganwadi_list_serializer(
-        AanganwadiCollection.find({"_id": ObjectId(id)}))
-    return {"status": "ok", "data": aanganwadi}
-
-
-@aanganwadi_router.put("/api/updateAanganwadi/{id}")
-async def update_aanganwadi(id: str, aanganwadi: Aanganwadi):
-    AanganwadiCollection.find_one_and_update({"_id": ObjectId(id)},
-                                             {"$set": dict(aanganwadi)})
-    aanganwadi = aanganwadi_list_serializer(
-        AanganwadiCollection.find({"_id": ObjectId(id)}))
-    return {"status": "ok", "data": aanganwadi}
-
-
-@aanganwadi_router.delete("/api/delete_aanganwadi/{id}")
-async def delete_aanganwadi(id: str):
-    AanganwadiCollection.find_one_and_delete({"_id": ObjectId(id)})
-    return {"status": "ok", "data": []}
 
 
 @child_router.post("/api/add_child")
