+++ main.py
@@ -2,21 +2,16 @@ import os
 
 import uvicorn
 from dotenv import load_dotenv
-
 from fastapi import FastAPI, Request, Form, Depends, Response
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 from fastapi.responses import HTMLResponse, FileResponse, RedirectResponse
-
 from Backend.routes.routes import user_router
-
 from Backend.routes.routes import aaganwadi_summary
-from Backend.routes.routes import aanganwadi_router, child_router, child_malnutrition, supplement_details, \
-    program_joining
+from Backend.routes.routes import child_router, child_malnutrition, supplement_details, program_joining
 from Backend.routes.ngo_routes import ngo_router
+from Backend.routes.aaganwadi_routes import aanganwadi_router
 from Backend.routes.routes import donor_router, supp_router, program_router
-
-
 from app.routes import sign_router
